requirements.txt
streamlit
opencv-python-headless
pytesseract
numpy
Pillow